adding type conversion

+++ devcontroller/misc/thread.py
@@ -48,7 +48,7 @@ class CountdownThread(StoppableThread):
 
     def do_execute(self):
         print("waiting %s seconds:" % self.t)
-        self.t = ceil(self.t)
+        self.t = int(ceil(self.t))
         while self.t:
             mins, secs = divmod(self.t, 60)
             sys.stdout.write('\rremaining ' + '{:02d}:{:02d}'.format(mins, secs))
